Monomer.Checkerboard

Description

This is a checkerboard container (grid of boxes with alternating background colors). Provided widgets are distributed from left to right, top to bottom. If the are more provided widgets than there are boxes in the grid then last widgets are ignored.

checkerboard 3 3 [filler, filler, label ":D"]

Re-exported modules

module Monomer.Checkerboard.CheckerboardCfg

Constructors

checkerboard

Arguments

:: (WidgetModel s, WidgetEvent e, Traversable t)

=> Int
Number of columns.

-> Int
Number of rows.

-> t (WidgetNode s e)
The list of items.

-> WidgetNode s e
The created checkerboard.

Creates a checkerboard container for multiple items using number of columns and rows.

checkerboard_

Arguments

:: (WidgetModel s, WidgetEvent e, Traversable t)

=> Int
Number of columns.

-> Int
Number of rows.

-> [CheckerboardCfg]
The config options.

-> t (WidgetNode s e)
The list of items.

-> WidgetNode s e
The created checkerboard.

Creates a checkerboard container for multiple items using number of columns and rows. Accepts config.