Monomer.Checkerboard.CheckerboardCfg

Configuration

data CheckerboardCfg

Configuration options for checkerboard:

lightColor: color of light squares in the checkerboard pattern.

darkColor: color of dark squares in the checkerboard pattern.

Constructors

CheckerboardCfg

Fields

_ccBgLightColor :: Maybe Color

_ccBgDarkColor :: Maybe Color

lightColor :: Color -> CheckerboardCfg

Sets color of light squares in the checkerboard pattern.

darkColor :: Color -> CheckerboardCfg

Sets color of dark squares in the checkerboard pattern.