Monomer.Dragboard.DragboardCfg

Re-exported modules

module Monomer.Checkerboard.CheckerboardCfg

Configuration

data DragboardCfg s e a

Configuration options for dragboard:

moveValidator: function to check validity of a move.

dragIdOffset: offset for drag and drop event messages.

selectColor: color of selected square.

disableClick: whether items can be moved only by dragging.

checkerConfig: config options for checkerboard container.

onFocus: event to raise when focus is received.

onFocusReq: WidgetRequest to generate when focus is received.

onBlur: event to raise when focus is lost.

onBlurReq: WidgetRequest to generate when focus is lost.

onChange: event to raise when the board changes.

onChangeReq: WidgetRequest to generate when the board changes.

Constructors

DragboardCfg

Fields

_dcValidator :: Maybe (Info a -> Bool)

_dcOffset :: Maybe Int

_dcSelectColor :: Maybe Color

_dcNoClick :: Maybe Bool

_dcCheckerCfg :: [CheckerboardCfg]

_dcOnFocusReq :: [Path -> WidgetRequest s e]

_dcOnBlurReq :: [Path -> WidgetRequest s e]

_dcOnChangeReq :: [Info a -> WidgetRequest s e]

moveValidator :: (Info a -> Bool) -> DragboardCfg s e a

Receives previous board, index of a square where item has been dragged to and index of a square where item has been dragged from and returns whether this move is valid or not. If move is not valid then the board state will not change.

dragIdOffset :: Int -> DragboardCfg s e a

When there are multiple dragboards, it is possible to drag an item from one dragboard to another. In order to ignore drop events caused by foreign items or to process them properly, the dragboards should use different offsets (for example, if each dragboard has less than 1000 squares, then multiples of 1000 can be used as offsets):

vgrid
    [ dragboard 3 3 boardState f
    , dragboard_ 3 3 anotherBoardState f [dragIdOffset 1000]
    , dragboard_ 3 3 yetAnotherBoardState f [dragIdOffset 2000]
    ]

selectColor :: Color -> DragboardCfg s e a

Color of selected square which is yellow by default.

disableClick :: DragboardCfg s e a

Allows items to be moved only by dragging.

disableClick_ :: Bool -> DragboardCfg s e a

Whether items can be moved only by dragging.

checkerConfig :: [CheckerboardCfg] -> DragboardCfg s e a

Config options for checkerboard container which is used by dragboard.