Monomer.EnhancedSlider

Description

This is a slider with a label, which shows current value, and buttons to increase and decrease value.

enhancedSlider lens 0 100

Re-exported modules

module Monomer.EnhancedSlider.EnhancedSliderCfg

Constructors

enhancedSlider

Arguments

:: (WidgetModel s, WidgetEvent e, SliderValue a)

=> ALens' s a
The lens into the model.

-> a
Minimum value.

-> a
Maximum value.

-> WidgetNode s e
The created enhanced slider.

Creates an enhanced slider using the given lens, providing minimum and maximum values.

enhancedSlider_

Arguments

:: (WidgetModel s, WidgetEvent e, SliderValue a)

=> ALens' s a
The lens into the model.

-> a
Minimum value.

-> a
Maximum value.

-> [EnhancedSliderCfg s e a]
The config options.

-> WidgetNode s e
The created enhanced slider.

Creates an enhanced slider using the given lens, providing minimum and maximum values. Accepts config.

enhancedSliderV

Arguments

:: (WidgetModel s, WidgetEvent e, SliderValue a)

=> a
The current value.

-> (a -> e)
The event to raise on change.

-> a
Minimum value.

-> a
Maximum value.

-> WidgetNode s e
The created enhanced slider.

Creates an enhanced slider using the given value and onChange event handler, providing minimum and maximum values.

enhancedSliderV_

Arguments

:: (WidgetModel s, WidgetEvent e, SliderValue a)

=> a
The current value.

-> (a -> e)
The event to raise on change.

-> a
Minimum value.

-> a
Maximum value.

-> [EnhancedSliderCfg s e a]
The config options.

-> WidgetNode s e
The created enhanced slider.

Creates an enhanced slider using the given value and onChange event handler, providing minimum and maximum values. Accepts config.

enhancedSliderD_

Arguments

:: (WidgetModel s, WidgetEvent e, SliderValue a)

=> WidgetData s a
The WidgetData to retrieve the value from.

-> a
Minimum value.

-> a
Maximum value.

-> [EnhancedSliderCfg s e a]
The config options.

-> [CompositeCfg a (EnhancedSliderEvent a) s e]
The composite config options.

-> WidgetNode s e
The created enhanced slider.

Creates an enhanced slider providing a WidgetData instance, minimum and maximum values and config.