Monomer.Graph

Description

This widget renders a Cartesian coordinate system and plots function graphs by connecting provided points. Coordinate system can be dragged and scaled. It is possible to render single points too.

This widget can receive GraphMsg messages:

GraphSetTranslation Point

GraphSetScale Point

GraphReset

graph [[(1,2), (1,3)], [(0,0), (1,1)]]
graphWithColors [(red, [(1,2), (1,3)]), (blue, [(0,0), (1,1)])]
graphWithData [[graphPoint (0, 0), graphColor red]]

Re-exported modules

module Monomer.Graph.GraphCfg

module Monomer.Graph.GraphData

module Monomer.Graph.GraphMsg

Constructors

graph

Arguments

:: [[(Double, Double)]]
The list with points.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with points.

graph_

Arguments

:: [[(Double, Double)]]
The list with points.

-> [GraphCfg]
The config options.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with points. Accepts config.

graphWithColors

Arguments

:: [(Color, [(Double, Double)])]
The list with colors and points.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with colors and points.

graphWithColors_

Arguments

:: [(Color, [(Double, Double)])]
The list with colors and points.

-> [GraphCfg]
The config options.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with colors and points. Accepts config.

graphWithData

Arguments

:: [[GraphData]]
The list with GraphData.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with GraphData.

graphWithData_

Arguments

:: [[GraphData]]
The list with GraphData.

-> [GraphCfg]
The config options.

-> WidgetNode s e
The created graph plotter.

Creates a graph plotter using the list with GraphData. Accepts config.