Monomer.Graph.GraphCfg

Configuration

data GraphCfg s e

Configuration options for graph:

wheelRate: speed of scaling.

limitX: limits along X-axis.

limitY: limits along Y-axis.

minimumX: left limit along X-axis.

maximumX: right limit along X-axis.

minimumY: bottom limit along Y-axis.

maximumY: top limit along Y-axis.

minScale: minimum scale along both X-axis and Y-axis.

maxScale: maximum scale along both X-axis and Y-axis.

minScaleX: minimum scale along X-axis.

maxScaleX: maximum scale along X-axis.

minScaleY: minimum scale along Y-axis.

maxScaleY: maximum scale along Y-axis.

lockX: lock X-axis (scale only Y-axis).

lockY: lock Y-axis (scale only X-axis).

hideMinorGridlines: whether to hide minor gridlines.

hideAxisNumbers: whether to hide axis numbers.

hideGrid: whether to hide all gridlines and axis numbers.

graphColors: which colors should be used to plot graphs.

onRightClick: event to raise on right click.

onRightClickReq: WidgetRequest to generate on right click.

Constructors

GraphCfg

Fields

_gcWheelRate :: Maybe Double

_gcMinX :: Maybe Double

_gcMaxX :: Maybe Double

_gcMinY :: Maybe Double

_gcMaxY :: Maybe Double

_gcMinScaleX :: Maybe Double

_gcMaxScaleX :: Maybe Double

_gcMinScaleY :: Maybe Double

_gcMaxScaleY :: Maybe Double

_gcLockX :: Maybe Bool

_gcLockY :: Maybe Bool

_gcHideMinor :: Maybe Bool

_gcHideNumbers :: Maybe Bool

_gcHideGrid :: Maybe Bool

_gcGraphColors :: Maybe [Color]

_gcOnRightClickReq :: [(Double, Double) -> WidgetRequest s e]

limitX :: (Double, Double) -> GraphCfg s e

Limits along X-axis.

limitY :: (Double, Double) -> GraphCfg s e

Limits along Y-axis.

minimumX :: Double -> GraphCfg s e

Left limit along X-axis.

maximumX :: Double -> GraphCfg s e

Right limit along X-axis.

minimumY :: Double -> GraphCfg s e

Bottom limit along Y-axis.

maximumY :: Double -> GraphCfg s e

Top limit along Y-axis.

minScale :: Double -> GraphCfg s e

Minimum scale along both X-axis and Y-axis.

maxScale :: Double -> GraphCfg s e

Maximum scale along both X-axis and Y-axis.

minScaleX :: Double -> GraphCfg s e

Minimum scale along X-axis.

maxScaleX :: Double -> GraphCfg s e

Maximum scale along X-axis.

minScaleY :: Double -> GraphCfg s e

Minimum scale along Y-axis.

maxScaleY :: Double -> GraphCfg s e

Maximum scale along Y-axis.

lockX :: GraphCfg s e

Lock X-axis (scale only Y-axis).

lockX_ :: Bool -> GraphCfg s e

Whether X-axis is locked and only Y-axis is scaled.

lockY :: GraphCfg s e

Lock Y-axis (scale only X-axis).

lockY_ :: Bool -> GraphCfg s e

Whether Y-axis is locked and only X-axis is scaled.

hideMinorGridlines :: GraphCfg s e

Hide minor gridlines.

hideMinorGridlines_ :: Bool -> GraphCfg s e

Whether to hide minor gridlines.

hideAxisNumbers :: GraphCfg s e

Hide axis numbers.

hideAxisNumbers_ :: Bool -> GraphCfg s e

Whether to hide axis numbers.

hideGrid :: GraphCfg s e

Hide all gridlines and axis numbers.

hideGrid_ :: Bool -> GraphCfg s e

Whether to hide all gridlines and axis numbers.

graphColors :: [Color] -> GraphCfg s e

List of colors which are used to plot graphs. This list is then cycled when plotting graphs (in case there are more graphs than provided colors).

onRightClick :: WidgetEvent e => ((Double, Double) -> e) -> GraphCfg s e

Event to raise on right click.

onRightClickReq :: ((Double, Double) -> WidgetRequest s e) -> GraphCfg s e

WidgetRequest to generate on right click.