Monomer.Graph.GraphData

Documentation

data GraphData

Options for graph data:

graphPoint: render single point.

graphPoints: use multiple points.

graphColor: set the color.

graphWidth: set the width of the line.

graphSeparate: whether the points should be rendered separately.

graphFill: whether to fill the area surrounded by points.

graphFillAlpha: transparency level of the filled area.

Constructors

GraphData

Fields

_gdPoints :: [(Double, Double)]

_gdColor :: Maybe Color

_gdWidth :: Maybe Double

_gdSeparate :: Maybe Bool

_gdFill :: Maybe Bool

_gdFillAlpha :: Maybe Double

graphPoint :: (Double, Double) -> GraphData

Render single point.

graphPoints :: [(Double, Double)] -> GraphData

Use multiple points.

graphColor :: Color -> GraphData

Set the color (if this option is not used then the graph will not be rendered).

graphWidth :: Double -> GraphData

Width of the line connecting provided points. If only single point is rendered then its radius will be twice the width.

graphSeparate :: GraphData

Do not connect the points and render them separately. Used when all points in the collection must have the same color.

graphSeparate_ :: Bool -> GraphData

Whether the points should be rendered separately.

graphFill :: GraphData

Fill the area surrounded by provided points with the color.

graphFill_ :: Bool -> GraphData

Whether to fill the area surrounded by provided points.

graphFillAlpha :: Double -> GraphData

Transparency level of the filled area.