Monomer.SaveManager

Description

This is a composite which consists of selectList and buttons:

New slot

Save

Load

Remove

saveManager field

Re-exported modules

module Monomer.SaveManager.SaveManagerModel

module Monomer.SaveManager.SaveManagerCfg

Configuration

data SaveManagerEvent a

Constructors

saveManager

Arguments

:: (WidgetModel s, WidgetEvent e, Eq a, Typeable a)

=> ALens' s (SaveManagerModel a)
The lens into the model.

-> WidgetNode s e
The created save manager.

Creates a save manager using the given lens.

saveManager_

Arguments

:: (WidgetModel s, WidgetEvent e, Eq a, Typeable a)

=> ALens' s (SaveManagerModel a)
The lens into the model.

-> [SaveManagerCfg s e a]
The config options.

-> WidgetNode s e
The created save manager.

Creates a save manager using the given lens. Accepts config.

saveManagerV

Arguments

:: (WidgetModel s, WidgetEvent e, Eq a, Typeable a)

=> SaveManagerModel a
The composite model.

-> (a -> e)
The event to raise on change.

-> WidgetNode s e
The created save manager.

Creates a save manager using the given composite model and onChange event handler.

saveManagerV_

Arguments

:: (WidgetModel s, WidgetEvent e, Eq a, Typeable a)

=> SaveManagerModel a
The composite model.

-> (a -> e)
The event to raise on change.

-> [SaveManagerCfg s e a]
The config options.

-> WidgetNode s e
The created save manager.

Creates a save manager using the given composite model and onChange event handler. Accepts config.

saveManagerD_

Arguments

:: (WidgetModel s, WidgetEvent e, Eq a, Typeable a)

=> WidgetData s (SaveManagerModel a)
The WidgetData to retrieve the model from.

-> [SaveManagerCfg s e a]
The config options.

-> [CompositeCfg (SaveManagerModel a) (SaveManagerEvent a) s e]
The composite config options.

-> WidgetNode s e
The created save manager.

Creates a save manager providing a WidgetData instance and config.