Monomer.EnhancedSlider.EnhancedSliderCfg

Configuration

data EnhancedSliderCfg s e a

Configuration options for enhancedSlider:

dragRate: the rate at which drag movement affects the number.

hideLabel: don't show the label, leave only slider and buttons.

titleCaption: the title for the shown value.

titleMethod: function to generate the label with value.

alignLeft: put horizontal slider to the left of the buttons. This is default.

alignCenter: put horizontal slider between the buttons.

alignRight: put horizontal slider to the right of the buttons.

alignTop: put vertical slider to the top of the buttons.

alignMiddle: put vertical slider between the buttons.

alignBottom: put vertical slider to the bottom of the buttons.

onFocus: event to raise when focus is received.

onFocusReq: WidgetRequest to generate when focus is received.

onBlur: event to raise when focus is lost.

onBlurReq: WidgetRequest to generate when focus is lost.

onChange: event to raise when the value changes.

onChangeReq: WidgetRequest to generate when the value changes.

Constructors

EnhancedSliderCfg

Fields

_escDragRate :: Maybe Rational

_escHideLabel :: Maybe Bool

_escTitle :: Maybe Text

_escTitleMethod :: Maybe (a -> Text)

_escAlignH :: Maybe AlignH

_escAlignV :: Maybe AlignV

_escOnFocusReq :: [Path -> WidgetRequest s e]

_escOnBlurReq :: [Path -> WidgetRequest s e]

_escOnChangeReq :: [a -> WidgetRequest s e]

titleMethod :: (a -> Text) -> EnhancedSliderCfg s e a

Receives function which converts the value into text and uses it to generate the label. Should be used if the title depends on the value or different formatting is needed.

hideLabel :: EnhancedSliderCfg s e a

Should be used when the label with the current value is not needed.

hideLabel_ :: Bool -> EnhancedSliderCfg s e a