Monomer.Graph.GraphCfg

Configuration

data GraphCfg

Configuration options for graph:

wheelRate: speed of scaling.

lockX: lock X-axis (scale only Y-axis).

lockY: lock Y-axis (scale only X-axis).

hideMinorGridlines: whether to hide minor gridlines.

hideAxisNumbers: whether to hide axis numbers.

graphColors: which colors should be used to plot graphs.

Constructors

GraphCfg

Fields

_gcWheelRate :: Maybe Double

_gcLockX :: Maybe Bool

_gcLockY :: Maybe Bool

_gcHideMinor :: Maybe Bool

_gcHideNumbers :: Maybe Bool

_gcGraphColors :: Maybe [Color]

lockX :: GraphCfg

Lock X-axis (scale only Y-axis).

lockX_ :: Bool -> GraphCfg

Whether X-axis is locked and only Y-axis is scaled.

lockY :: GraphCfg

Lock Y-axis (scale only X-axis).

lockY_ :: Bool -> GraphCfg

Whether Y-axis is locked and only X-axis is scaled.

hideMinorGridlines :: GraphCfg

Hide minor gridlines.

hideMinorGridlines_ :: Bool -> GraphCfg

Whether to hide minor gridlines.

hideAxisNumbers :: GraphCfg

Hide axis numbers.

hideAxisNumbers_ :: Bool -> GraphCfg

Whether to hide axis numbers.

graphColors :: [Color] -> GraphCfg

List of colors which are used to plot graphs. This list is then cycled when plotting graphs (in case there are more graphs than provided colors).