Monomer.Graph.GraphData

Documentation

data GraphData s e

Options for graph data:

graphPoint: render single point.

graphPoints: use multiple points.

graphColor: set the color.

graphHoverColor: set the color of hovered point.

graphActiveColor: set the color of dragged point.

graphWidth: set the width of the line.

graphRadius: set the radius of the points.

graphSeparate: whether the points should be rendered separately.

graphFill: whether to fill the area surrounded by points.

graphFillAlpha: transparency level of the filled area.

graphOnChange: event to raise when a point is dragged.

graphOnChangeReq: WidgetRequest to generate when a point is dragged.

graphOnClick: event to raise when a point is clicked.

graphOnClickReq: WidgetRequest to generate when a point is clicked.

Constructors

GraphData

Fields

_gdPoints :: [(Double, Double)]

_gdColor :: Maybe Color

_gdHoverColor :: Maybe Color

_gdActiveColor :: Maybe Color

_gdWidth :: Maybe Double

_gdRadius :: Maybe Double

_gdSeparate :: Maybe Bool

_gdFill :: Maybe Bool

_gdFillAlpha :: Maybe Double

_gdChangeReq :: [Int -> (Double, Double) -> WidgetRequest s e]

_gdClickReq :: [Int -> WidgetRequest s e]

graphPoint :: (Double, Double) -> GraphData s e

Render single point.

graphPoints :: [(Double, Double)] -> GraphData s e

Use multiple points.

graphColor :: Color -> GraphData s e

Set the color (if this option is not used then the graph will not be rendered).

graphHoverColor :: Color -> GraphData s e

Set the color of hovered point (if this option is not used then the color set by graphColor is used).

graphActiveColor :: Color -> GraphData s e

Set the color of dragged point (if this option is not used then the color set by graphColor is used).

graphWidth :: Double -> GraphData s e

Width of the line connecting provided points. If only single point is rendered then its radius will be twice the width.

graphRadius :: Double -> GraphData s e

Radius of the provided points if they are rendered separately. The difference between graphWidth and graphRadius is that the former is given in pixels and renders the same when the scale changes while the latter is given in units of the Cartesian coordinate system and hence the points become bigger or smaller when the scale changes.

graphSeparate :: GraphData s e

Do not connect the points and render them separately. Used when all points in the collection must have the same color.

graphSeparate_ :: Bool -> GraphData s e

Whether the points should be rendered separately.

graphFill :: GraphData s e

Fill the area surrounded by provided points with the color.

graphFill_ :: Bool -> GraphData s e

Whether to fill the area surrounded by provided points.

graphFillAlpha :: Double -> GraphData s e

Transparency level of the filled area.

graphOnChange :: WidgetEvent e => (Int -> (Double, Double) -> e) -> GraphData s e

Raises an event when a point is dragged by passing its index and new coordinates. This option is ignored if graphSeparate is not enabled.

graphOnChangeReq :: (Int -> (Double, Double) -> WidgetRequest s e) -> GraphData s e

Generates a WidgetRequest when a point is dragged by passing its index and new coordinates. This option is ignored if graphSeparate is not enabled.

graphOnClick :: WidgetEvent e => (Int -> e) -> GraphData s e

Raises an event when a point is clicked by passing its index. This option is ignored if graphSeparate is not enabled.

graphOnClickReq :: (Int -> WidgetRequest s e) -> GraphData s e

Generates a WidgetRequest when a point is clicked by passing its index. This option is ignored if graphSeparate is not enabled.